send

+++ main.py
@@ -171,14 +171,14 @@ async def predict(file: UploadFile = File(...)):
     try:
         contents = await file.read()
         image = Image.open(io.BytesIO(contents)).convert("RGB")
-        image = image.resize((126, 126))  # ✅ resize to the correct shape
+        image = image.resize((126, 126))  # use the correct size expected by your model
         image_array = np.array(image) / 255.0
-        image_array = image_array.reshape(1, 126, 126, 3)  # ✅ match model's input shape
+        image_array = image_array.reshape(1, 126, 126, 3)
 
         predictions = model.predict(image_array)
         predicted_class_index = np.argmax(predictions[0])
         predicted_class = class_names[predicted_class_index]
-        confidence = round(100 * np.max(predictions[0]), 2)
+        confidence = float(round(100 * np.max(predictions[0]), 2))  # ✅ cast to native float
         treatment = get_cure(predicted_class)
 
         return JSONResponse(content={
